Refactor Streamlit tabs into separate modules
